Revert "LPD-79062 moves relevant classes from feature-flag-web to portal-impl"
This reverts commit 0e22ae3b72c0ae98480484f5f5fdf97a9c71343e.
