+++ modules/apps/feature-flag/feature-flag-web/src/main/java/com/liferay/feature/flag/web/internal/feature/flag/FeatureFlagsBag.java
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.feature.flag;
+package com.liferay.feature.flag.web.internal.feature.flag;
 
+import com.liferay.feature.flag.web.internal.model.FeatureFlagWrapper;
+import com.liferay.feature.flag.web.internal.model.PreferenceAwareFeatureFlag;
 import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.feature.flag.model.FeatureFlagWrapper;
-import com.liferay.portal.feature.flag.model.PreferenceAwareFeatureFlag;
 import com.liferay.portal.kernel.feature.flag.FeatureFlag;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
