+++ portal-impl/src/META-INF/portal-spring.xml
@@ -391,7 +391,4 @@
 	<bean class="com.liferay.portlet.social.service.persistence.impl.SocialRequestModelArgumentsResolver" id="com.liferay.portlet.social.service.persistence.impl.SocialRequestModelArgumentsResolver" />
 	<bean class="com.liferay.portlet.social.service.persistence.impl.SocialRequestPersistenceImpl" id="com.liferay.social.kernel.service.persistence.SocialRequestPersistence" />
 	<bean class="com.liferay.portlet.social.service.impl.SocialRequestInterpreterLocalServiceImpl" id="com.liferay.social.kernel.service.SocialRequestInterpreterLocalService" />
-	<bean class="com.liferay.portal.feature.flag.FeatureFlagManagerImpl" id="com.liferay.portal.kernel.feature.flag.FeatureFlagManager" />
-	<bean class="com.liferay.portal.feature.flag.FeatureFlagsBagProviderImpl" id="com.liferay.portal.feature.flag.FeatureFlagsBagProvider" />
-	<bean class="com.liferay.portal.feature.flag.manager.FeatureFlagPreferencesManager" id="com.liferay.portal.feature.flag.manager.FeatureFlagPreferencesManager" />
 </beans>
